enhance charts chapter for the tutorial
